Move Network class to src/dependencies (fixes #1349 )
